moved get_all_generators to utils from tests as this is useful
