+++ pynames/__init__.py
@@ -7,3 +7,4 @@ import pynames.korean
 import pynames.russian
 
 from pynames.relations import GENDER, LANGUAGE
+from pynames.utils import get_all_generators
